+++ lua/user/core/init.lua
@@ -1,0 +1,2 @@
+require "user.core.keymaps"
+require "user.core.options"
