+++ lua/user/plugins/lsp/init.lua
@@ -1,0 +1,8 @@
+local status_ok, _ = pcall(require, "lspconfig")
+if not status_ok then
+  return
+end
+
+require "user.plugins.lsp.mason"
+require("user.plugins.lsp.handlers").setup()
+require "user.plugins.lsp.null-ls"
