--- lua/user/plugins_config/init.lua
@@ -1,21 +1,0 @@
-require "user.plugins_config.colorscheme"
-require "user.plugins_config.cmp"
-require "user.plugins_config.telescope"
-require "user.plugins_config.gitsigns"
-require "user.plugins_config.git-blame"
-require "user.plugins_config.treesitter"
-require "user.plugins_config.autopairs"
-require "user.plugins_config.comment"
-require "user.plugins_config.nvim-tree"
-require "user.plugins_config.bufferline"
-require "user.plugins_config.lualine"
-require "user.plugins_config.toggleterm"
-require "user.plugins_config.project"
-require "user.plugins_config.impatient"
-require "user.plugins_config.indentline"
-require "user.plugins_config.alpha"
-require "user.plugins_config.whichkey"
-require "user.plugins_config.autocommands"
-require "user.plugins_config.rust_config"
-require "user.plugins_config.dap_config"
-require "user.plugins_config.vim-rest-console"
